requirements.txt
flask
python-dotenv
requests
beautifulsoup4
anthropic
tavily-python
